pyproject.toml
[project]
name = "nonebot-plugin-mcsm"
version = "0.2.0"
description = "nonebot-plugin-mcsm"
readme = "README.md"
requires-python = ">=3.8, <4.0"

[tool.nonebot]
plugins = ["nonebot_plugin_mcsm"]

[tool.poetry]
name = "nonebot-plugin-mcsm"
version = "0.2.0"
description = "MCSM plugin for NoneBot2"
authors = ["LiLuo-B"]
license = "GPLv3"
readme = "README.md"
repository = "https://github.com/LiLuo-B/nonebot-plugin-mcsm"
documentation = "https://github.com/LiLuo-B/nonebot-plugin-mcsm/blob/main/README.md"
[tool.poetry.dependencies]
python = ">=3.8,<4.0"
nonebot2 = ">=2.2.0"
nonebot-plugin-alconna = ">=0.46.6"
nonebot-plugin-htmlrender = ">=0.3.1"
httpx = ">=0.13.0"
jinja2 = ">=3.1.4"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
